refactor(sse): extract SSE parsing utilities into sse/ module
Spec 2.2 complete: Extract SSE parsing functions from provider-utils.ts
into dedicated sse/ module.

Create sse/parsing.ts with:
- extractSseDataLines (internal)
- stripSseCommentLines
- chatCompletionHasReasoningContent
- streamPayloadHasReasoningTrace
- streamPayloadHasSubstantiveChunks

Update imports in:
- provider-strategy/base.ts
- provider-strategy/strategies/standard.ts
- tests/reasoning-stream-check.test.ts

Keep streamPayloadIndicatesQuotaError in provider-utils.ts since it
depends on error classification functions (to be extracted in next spec).

All 579 tests pass.
