+++ src/lib/provider-strategy/base.ts
@@ -5,12 +5,14 @@ import type { FastifyReply } from "fastify";
 import { copyUpstreamHeaders } from "../proxy.js";
 import { chatCompletionToSse } from "../responses-compat.js";
 import {
+  streamPayloadHasReasoningTrace,
+  stripSseCommentLines,
+  streamPayloadHasSubstantiveChunks,
   chatCompletionHasReasoningContent,
+} from "../sse/index.js";
+import {
   responseIndicatesMissingModel,
   responseIndicatesModelNotSupportedForAccount,
-  stripSseCommentLines,
-  streamPayloadHasReasoningTrace,
-  streamPayloadHasSubstantiveChunks,
   streamPayloadIndicatesQuotaError,
   summarizeUpstreamError,
 } from "../provider-utils.js";
