+++ src/lib/sse/index.ts
@@ -1,0 +1,7 @@
+export {
+  extractSseDataLines,
+  stripSseCommentLines,
+  chatCompletionHasReasoningContent,
+  streamPayloadHasReasoningTrace,
+  streamPayloadHasSubstantiveChunks,
+} from "./parsing.js";
