+++ src/tests/reasoning-stream-check.test.ts
@@ -1,10 +1,7 @@
 import assert from "node:assert/strict";
 import test from "node:test";
 
-import {
-  streamPayloadHasReasoningTrace,
-  chatCompletionHasReasoningContent,
-} from "../lib/provider-utils.js";
+import { streamPayloadHasReasoningTrace, chatCompletionHasReasoningContent } from "../lib/sse/index.js";
 
 test("streamPayloadHasReasoningTrace detects chat completions reasoning_content", () => {
   const payload = [
